# Key embedded in dictEntry - CPU cache friendly