Whenever a key is looked up, the dictEntry (hashtable bucket) and the key
are accessed together. This PR embeds the key in the same allocation as the
dictEntry, which makes sure they are in the same cache line, thus saving
one slow memory access.

This compensates for the potential regression from #9356 for key lengths
15-22 in cluster mode. Furthermore, latency is expected to improve for
other key lengths in cluster mode as well as in plain Redis. Benchmarks
will show.

The callbacks in dictType are changed in the following ways:

* dictEntryMetadataBytes is replaced by createEntry, which (if present)
  takes care of allocating the entry and setting the key.
* keyDestructor and valueDestructor are passed the dictEntry as an extra
  argument, so they can check if the key is embedded in the dictEntry and
  then not try to free it.
* repairEntry is added, which is called from active defrag code when an
  entry has been reallocated, moving custom logic from the active defrag
  code to the dict type.

TODO:

* [x] Working implementation. Tests passing.
* [x] Benchmarks for cluster/plain Redis, key lengths of different
      allocation sizes.
* [x] Memory usage comparison.

Open questions:

* What's the maximum key length we want to embed in a dictEntry?
* Do we want a similar optimization for other dicts, e.g. hashes?