@madolson 2021-09-06
It's unfortunate that we're paying 8 bytes for the key ptr even though it doesn't do anything. We could stuff a bit in the bottom 3 bits of 8 bit aligned systems, since their pointers always end in 000, which could be used to indicate if it's the new address/old address scheme. We would need to go through and make sure everyone only uses deGetKey, which would handle the pointer maintenance. I'm not sure if that would be too much of over optimizing though.