@oranagra 2021-09-07
I'm concerned that we may be going too far towards coupling more and more parts of Redis with that dict, which we also wanna replace one day (due to the rehashing memory issue) 
It's no reason not to do incremental changes, but I'm just stressing that when the day comes we may have more areas in the code that are tightly coupled and will need more work. 
That said, let's proceed with this task and see where it leads us, and make the call when ready.