@zuiderkwast 2021-09-07
> It's unfortunate that we're paying 8 bytes for the key ptr even though it doesn't do anything. We could stuff a bit in the bottom 3 bits of 8 bit aligned systems, since their pointers always end in 000, which could be used to indicate if it's the new address/old address scheme. 

@madolson An sds string is an odd pointer, since its header (an odd number of bytes) is before the pointer. There might still be bits to use though, for example in the 'value' or the 'next' pointer. (We'd need to move around the dictEntry fields, since the embedded key needs to be last.) We have the same waste of bits in the robj with embedded sds (many DB values). I think it can be a different PR.

> I'm concerned that we may be going too far towards coupling more and more parts of Redis with that dict, which we also wanna replace one day (due to the rehashing memory issue)

@oranagra I'm well aware. My #8700 was stalled since I couldn't make the build pass, but now it's diverged when #9356 was merged. It's still possible to rebase the HAMT. We'd need to wrap the key-with-cluster-metadata-and-maybe-embedded-key in another struct (basically the dictEntry without the value and next-pointer).

The build's passing now, so it's ready for (pre-)review and benchmarks.