@zuiderkwast 2021-09-07
Memory usage analysis

<details>
  <summary>Details (TL;DR Fewer allocations, no impact on total usage)</summary>

We have the following allocator bin sizes: 8 16 24 32 40 48 56 64
80 96 112 128 160 192 224 256 ...

In non-cluster mode, the dictEntry is 24 bytes and in cluster mode, it's 40 bytes.

Keys up to length 31 have 2 bytes overhead (SDS type 5) and lengths 32-255 have 4 bytes overhead (SDS type 8). We only embed sds type 5 keys and only if the combined size is up to 64 bytes.

| Key size | non-cluster, separate allocations | non-cluster, combined allocation  |
|----------|----------------------|----------------------|
| 1-6      | 8 + 24 = 32          | 32                   |
| 7-14     | 16 + 24 = 40         | 40                   |
| 15-22    | 24 + 24 = 48         | 48                   |
| 23-30    | 32 + 24 = 56         | 56                   |
| 31       | 40 + 24 = 64         | 64                   |
| 32-36    | 40 + 24 = 64         | 64 (not used)        |

| Key size | cluster, separate allocations | cluster, combined allocation |
|----------|----------------------|----------------------|
| 1-6      | 8 + 40 = 48          | 48                   |
| 7-14     | 16 + 40 = 56         | 56                   |
| 15-22    | 24 + 40 = 64         | 64                   |
| 23-30    | 32 + 40 = 72         | 80 (not used)        |

As seen above, the total allocation size is the same for combined and separate allocations, for the cases where we use embedded keys.
</details>