@zuiderkwast 2021-09-07
Benchmark: 14 bytes keys, integer values, non-cluster (on laptop)

<details>
<summary>TL;DR 10% latency win!</summary>

This PR:

```
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q set k:__rand_int__ 1__rand_int__
set k:__rand_int__ 1__rand_int__: 908843.00 requests per second, p50=0.479 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q set k:__rand_int__ 1__rand_int__
set k:__rand_int__ 1__rand_int__: 1025325.50 requests per second, p50=0.431 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q set k:__rand_int__ 1__rand_int__
set k:__rand_int__ 1__rand_int__: 973899.50 requests per second, p50=0.455 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q get k:__rand_int__
get k:__rand_int__: 1210067.75 requests per second, p50=0.375 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q del k:__rand_int__
del k:__rand_int__: 1140901.25 requests per second, p50=0.383 msec
```

Unstable:

```
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q set k:__rand_int__ 1__rand_int
set k:__rand_int__ 1__rand_int: 815793.75 requests per second, p50=0.503 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q set k:__rand_int__ 1__rand_int
set k:__rand_int__ 1__rand_int: 814730.31 requests per second, p50=0.535 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q set k:__rand_int__ 1__rand_int
set k:__rand_int__ 1__rand_int: 867980.19 requests per second, p50=0.519 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q get k:__rand_int__
get k:__rand_int__: 975419.44 requests per second, p50=0.463 msec
$ ./redis-benchmark --threads 2 -P 10 -r 10000000 -n 10000000 -q del k:__rand_int__
```

</details>