@zuiderkwast 2021-09-08
Cluster Benchmark: 22 bytes keys, integer values (on laptop)

<details>
<summary>This is the one with a regression in #9356</summary>

Details below.

![image](https://user-images.githubusercontent.com/273886/132481964-71a3b15c-c232-42d1-83c1-f156304ae5da.png)

Explanation: In this case, "old unstable" has 24 bytes dictEntry and 24 bytes key allocation. The later ones have 40+24 or 64 bytes allocation(s).

Cluster: 3 masters, no replicas, no AOF, no rdb.

```
$ OPTS='--cluster -p 30001 --threads 2 -P 10 -r 10000000 -n 10000000 -q'
```

This PR:

```
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1813894.38 requests per second, p50=0.143 msec
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1736111.00 requests per second, p50=0.151 msec
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1814552.62 requests per second, p50=0.143 msec
$ ./redis-benchmark $OPTS get 'key:{tag}:__rand_int__'
get key:{tag}:__rand_int__: 1900779.25 requests per second, p50=0.143 msec
$ ./redis-benchmark $OPTS del 'key:{tag}:__rand_int__'
del key:{tag}:__rand_int__: 1904036.62 requests per second, p50=0.135 msec
```

Unstable ("dict metadata")

```
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1477759.62 requests per second, p50=0.295 msec
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1428163.38 requests per second, p50=0.295 msec
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1595659.75 requests per second, p50=0.279 msec
$ ./redis-benchmark $OPTS get 'key:{tag}:__rand_int__'
get key:{tag}:__rand_int__: 1899335.25 requests per second, p50=0.223 msec
$ ./redis-benchmark $OPTS del 'key:{tag}:__rand_int__'
del key:{tag}:__rand_int__: 1734605.38 requests per second, p50=0.263 msec
```

Old unstable (just before #9356 was merged):

```
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1329787.25 requests per second, p50=0.335 msec
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1330141.00 requests per second, p50=0.343 msec
$ ./redis-benchmark $OPTS set 'key:{tag}:__rand_int__' 1__rand_int__
set key:{tag}:__rand_int__ 1__rand_int__: 1477978.12 requests per second, p50=0.311 msec
$ ./redis-benchmark $OPTS get 'key:{tag}:__rand_int__'
get key:{tag}:__rand_int__: 1994813.50 requests per second, p50=0.215 msec
$ ./redis-benchmark $OPTS del 'key:{tag}:__rand_int__'
del key:{tag}:__rand_int__: 1425110.38 requests per second, p50=0.319 msec
```
</details>